added some JobEndpoint Tests:
