+++ src/test/java/uk/gov/ons/census/bulkprocessor/endpoint/JobEndPointIT.java
@@ -6,37 +6,23 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.FileSystemResource;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 import uk.gov.ons.census.bulkprocessor.model.dto.JobDto;
-import uk.gov.ons.census.bulkprocessor.model.dto.JobStatusDto;
 import uk.gov.ons.census.bulkprocessor.model.entity.BulkProcess;
+import uk.gov.ons.census.bulkprocessor.model.entity.Job;
+import uk.gov.ons.census.bulkprocessor.model.entity.JobStatus;
 import uk.gov.ons.census.bulkprocessor.model.repository.JobRepository;
-import uk.gov.ons.census.bulkprocessor.model.repository.JobRowRepository;
-import uk.gov.ons.census.bulkprocessor.models.test_dtos.CollectionCase;
-import uk.gov.ons.census.bulkprocessor.models.test_dtos.ResponseManagementEvent;
-import uk.gov.ons.census.bulkprocessor.testutils.QueueSpy;
-import uk.gov.ons.census.bulkprocessor.testutils.RabbitQueueHelper;
 
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static uk.gov.ons.census.bulkprocessor.model.entity.BulkProcess.NEW_ADDRESS;
 
 @ContextConfiguration
 @ActiveProfiles("test")
@@ -48,15 +34,11 @@ public class JobEndPointIT {
 
     @Autowired
     JobRepository jobRepository;
+//
+//    @Autowired
+//    JobRowRepository jobRowRepository;
 
-    @Autowired
-    JobRowRepository jobRowRepository;
-
-    @Autowired
-    private RabbitQueueHelper rabbitQueueHelper;
-
-    private String outputQueue = "case.rh.case";
-    String fileUploadUrl = "";
+    String jobUrl = "";
 
     RestTemplate restTemplate = new RestTemplate();
 
@@ -65,108 +47,38 @@ public class JobEndPointIT {
 
     @Before
     public void setUp() {
-        fileUploadUrl = "http://localhost:" + port + "/upload/";
-        rabbitQueueHelper.purgeQueue(outputQueue);
+        jobUrl = "http://localhost:" + port + "/job";
         jobRepository.deleteAll();
     }
 
 
     @Test
-    public void UploadANewAddressFile() throws Exception {
-
-        try (QueueSpy outputQueueSpy = rabbitQueueHelper.listen(outputQueue)) {
-            HttpEntity<MultiValueMap<String, Object>> requestEntity =
-                    buildEntityToSendFile(NEW_ADDRESS_FILE, NEW_ADDRESS);
-            ResponseEntity<String> response =
-                    restTemplate.postForEntity(fileUploadUrl, requestEntity, String.class);
-            assertThat(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
-
-            ResponseManagementEvent emittedCase = outputQueueSpy.checkExpectedMessageReceived();
-            assertThat(emittedCase.getPayload().getCollectionCase().getAddress().getUprn()).isEqualTo("6034022");
-            assertThat(emittedCase.getPayload().getCollectionCase().getAddress().getOrganisationName()).isEqualTo("TEAM RM");
-
-            JobDto job = getSingleJobAndCheckTheresOnlyOne();
-            assertThat(job.getJobStatus()).isEqualTo(JobStatusDto.PROCESSED_OK);
-
-            assertThat(job.getStagedRowCount()).isEqualTo(1);
-            assertThat(job.getFileName()).isEqualTo(NEW_ADDRESS_FILE);
-            assertThat(job.getBulkProcess()).isEqualTo("NEW_ADDRESS");
-            assertThat(job.getFileRowCount()).isEqualTo(2);
+    public void findAllJobsIT()  {
+        for( int i = 0; i < 10; i++) {
+            Job job = new Job();
+            job.setId(UUID.randomUUID());
+            job.setBulkProcess(BulkProcess.REFUSAL);
+            job.setJobStatus(JobStatus.PROCESSED_OK);
+            jobRepository.saveAndFlush(job);
         }
+
+        ResponseEntity<JobDto[]> actualJobsResponse = restTemplate.getForEntity(jobUrl, JobDto[].class);
+        assertThat(actualJobsResponse.getStatusCode()).isEqualTo(HttpStatus.OK);
+        JobDto[] actualJobs = actualJobsResponse.getBody();
+        assertThat(actualJobs.length).isEqualTo(10);
     }
 
     @Test
-    public void UploadARefusalFile() throws InterruptedException, Exception {
-        CollectionCase emittedCase;
+    public void testGetJobById() {
+        Job job = new Job();
+        job.setId(UUID.randomUUID());
+        job.setBulkProcess(BulkProcess.REFUSAL);
+        job.setJobStatus(JobStatus.PROCESSED_OK);
 
-        // 1st load a New Address to get a case to refuse
-        try (QueueSpy outputQueueSpy = rabbitQueueHelper.listen(outputQueue)) {
-            HttpEntity<MultiValueMap<String, Object>> requestEntity =
-                    buildEntityToSendFile(NEW_ADDRESS_FILE, NEW_ADDRESS);
+        jobRepository.saveAndFlush(job);
 
-            ResponseEntity<String> response =
-                    restTemplate.postForEntity(fileUploadUrl, requestEntity, String.class);
-            assertThat(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
-
-            emittedCase = outputQueueSpy.checkExpectedMessageReceived().getPayload().getCollectionCase();
-            assertThat(emittedCase.getAddress().getUprn()).isEqualTo("6034022");
-
-            JobDto job = getSingleJobAndCheckTheresOnlyOne();
-            assertThat(job.getJobStatus()).isEqualTo(JobStatusDto.PROCESSED_OK);
-        }
-
-        try (QueueSpy outtyQueue = rabbitQueueHelper.listen(outputQueue)) {
-            HttpEntity<MultiValueMap<String, Object>> entityToSend = buildEntityForRefusal(emittedCase.getId());
-            ResponseEntity<String> response = restTemplate.postForEntity(fileUploadUrl, entityToSend, String.class);
-            assertThat(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
-
-            CollectionCase refusalEmittedCaseUpdate = outtyQueue.checkExpectedMessageReceived().getPayload().getCollectionCase();
-            assertThat(refusalEmittedCaseUpdate.getId()).isEqualTo(emittedCase.getId());
-
-            assertThat(refusalEmittedCaseUpdate.getRefusalReceived().name()).isEqualTo("EXTRAORDINARY_REFUSAL");
-        }
-    }
-
-    private JobDto getSingleJobAndCheckTheresOnlyOne() {
-        String allJobsUrl = "http://localhost:" + port + "/job";
-        RestTemplate restTemplate = new RestTemplate();
-        ResponseEntity<JobDto[]> allJobsResponse =
-                restTemplate.getForEntity(allJobsUrl, JobDto[].class);
-        assertThat(allJobsResponse.getStatusCode()).isEqualTo(HttpStatus.OK);
-
-        JobDto[] allJobs = allJobsResponse.getBody();
-        assertThat(allJobs.length).isEqualTo(1);
-
-        return allJobs[0];
-    }
-
-
-    private HttpEntity<MultiValueMap<String, Object>> buildEntityForRefusal(UUID caseId) throws IOException {
-        String fileData = "case_id,refusal_type\n" + caseId + ",EXTRAORDINARY_REFUSAL";
-        Path tempFile = Files.createTempFile(null, null);
-        System.out.println(tempFile);
-
-        Files.write(tempFile, fileData.getBytes(StandardCharsets.UTF_8));
-        FileSystemResource fileSystemResource = new FileSystemResource(tempFile);
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.MULTIPART_FORM_DATA);
-
-        MultiValueMap<String, Object> body = new LinkedMultiValueMap<>();
-        body.add("file", fileSystemResource);
-        body.add("bulkProcess", BulkProcess.REFUSAL.toString());
-
-        return new HttpEntity<>(body, headers);
-    }
-
-    private HttpEntity<MultiValueMap<String, Object>> buildEntityToSendFile(
-            String fileName, BulkProcess bulkProcess) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.MULTIPART_FORM_DATA);
-        MultiValueMap<String, Object> body = new LinkedMultiValueMap<>();
-        body.add("file", new ClassPathResource(TEST_FILES_DIR + fileName));
-        body.add("bulkProcess", bulkProcess.toString());
-
-        return new HttpEntity<>(body, headers);
+        ResponseEntity<JobDto> actualJobResponse = restTemplate.getForEntity(jobUrl + "/" + job.getId().toString(), JobDto.class);
+        assertThat(actualJobResponse.getStatusCode()).isEqualTo(HttpStatus.OK);
+        assertThat(actualJobResponse.getBody().getId()).isEqualTo(job.getId());
     }
 }
